+++ src/models/likes.py
@@ -1,11 +1,11 @@
-from sqlalchemy import BigInteger, ForeignKey, UniqueConstraint
+from sqlalchemy import Integer, ForeignKey, UniqueConstraint
 
 from .abstract_model import AbstractModel
 from app_singleton import db
 
 class Likes (AbstractModel):
-    user_id = db.Column(BigInteger, ForeignKey("user.id"), nullable=False)
-    feed_id = db.Column(BigInteger, ForeignKey("feed.id"), nullable=False)
+    user_id = db.Column(Integer, ForeignKey("user.id"), nullable=False)
+    feed_id = db.Column(Integer, ForeignKey("feed.id"), nullable=False)
     
     user = db.relationship("User", foreign_keys=[user_id])
     feed = db.relationship("Feed", foreign_keys=[feed_id])
